% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ellipse.general.R
\name{ellipse.general}
\alias{ellipse.general}
\title{Draw a confidence ellipse using R}
\usage{
ellipse.general(mu, s, c)
}
\arguments{
\item{mu}{the mean vector}

\item{s}{the covariance matrix}

\item{c}{a constant}
}
\value{
some points of the confidence ellipse
}
\description{
Draw a confidence ellipse using R
}
\examples{
\dontrun{
    mu <- c(2.5, 2.5)  # 均值向量
    s <- matrix(c(2, 1, 1, 2), ncol = 2)  # 协方差矩阵
    c <- 1
    X <- ellipse.general(mu = mu, s = s, c = c)
}
}
