% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{maxcol}
\alias{maxcol}
\title{Calculate the maximum module length of each column in the matrix}
\usage{
maxcol(x)
}
\arguments{
\item{x}{a matrix}
}
\value{
a constant of maximum module length of each column in the matrix
}
\description{
Calculate the maximum module length of each column in the matrix
}
\examples{
\dontrun{
mu <- c(10,10)
s <- matrix(c(1,2,2,1,1,3,4,5,6,7,8,9), nrow = 2, ncol = 6)
n <- maxcol(s)
n
}
}
